Updates to local observability
Enabled a Search Agent to repeatedly created and solve new problems each time it is asked to provide an action. This allows the agent to respond to a changing environment. However an issue still remains in that the created problem is not persistent across iterations, so the agent forgets about walls it cannot currently see. Some method of maintaining wall discovery information from one problem to the next should solve this.